components arranged in separate folder
