+++ 05-eat-n-split/src/components/button.js
@@ -1,0 +1,7 @@
+export default function Button({ children, onClick }) {
+  return (
+    <button className="button" onClick={onClick}>
+      {children}
+    </button>
+  );
+}
