+++ 05-eat-n-split/src/components/friend-list.js
@@ -1,0 +1,15 @@
+import Friend from "./friend";
+export default function FriendsList({ friends, onSelection, selectedFriend }) {
+  return (
+    <ul>
+      {friends.map((friend) => (
+        <Friend
+          friend={friend}
+          onSelection={onSelection}
+          selectedFriend={selectedFriend}
+          key={friend.id}
+        />
+      ))}
+    </ul>
+  );
+}
