+++ 05-eat-n-split/src/components/friend.js
@@ -1,0 +1,29 @@
+import Button from "./button";
+export default function Friend({ friend, onSelection, selectedFriend }) {
+  const isSelected = selectedFriend?.id === friend.id;
+  return (
+    <li className={isSelected ? "selected" : ""}>
+      <img src={friend.image} alt="friend" />
+      <h3>{friend.name}</h3>
+
+      {friend.balance < 0 && (
+        <p className="red">
+          You owe {friend.name} {Math.abs(friend.balance)}
+          {"$ "}
+        </p>
+      )}
+
+      {friend.balance > 0 && (
+        <p className="green">
+          {friend.name} owes you {friend.balance}$
+        </p>
+      )}
+
+      {friend.balance === 0 && <p>{friend.name} and you are even.</p>}
+
+      <Button onClick={() => onSelection(friend)}>
+        {isSelected ? "Close" : "Select"}
+      </Button>
+    </li>
+  );
+}
